chore: clean up interactions types
