+++ payloads/v9/_interactions/responses.ts
@@ -1,0 +1,78 @@
+import { MessageFlags } from '../index';
+import { RESTPostAPIWebhookWithTokenJSONBody } from '../../../v9';
+
+/**
+ * https://discord.com/developers/docs/interactions/slash-commands#interaction-interactiontype
+ */
+export const enum InteractionType {
+	Ping = 1,
+	ApplicationCommand,
+	MessageComponent,
+}
+
+/**
+ * https://discord.com/developers/docs/interactions/slash-commands#interaction-response
+ */
+export type APIInteractionResponse =
+	| APIInteractionResponsePong
+	| APIInteractionResponseChannelMessageWithSource
+	| APIInteractionResponseDeferredChannelMessageWithSource
+	| APIInteractionResponseDeferredMessageUpdate
+	| APIInteractionResponseUpdateMessage;
+
+export interface APIInteractionResponsePong {
+	type: InteractionResponseType.Pong;
+}
+
+export interface APIInteractionResponseChannelMessageWithSource {
+	type: InteractionResponseType.ChannelMessageWithSource;
+	data: APIInteractionApplicationCommandCallbackData;
+}
+
+export interface APIInteractionResponseDeferredChannelMessageWithSource {
+	type: InteractionResponseType.DeferredChannelMessageWithSource;
+	data?: Pick<APIInteractionApplicationCommandCallbackData, 'flags'>;
+}
+
+export interface APIInteractionResponseDeferredMessageUpdate {
+	type: InteractionResponseType.DeferredMessageUpdate;
+}
+
+export interface APIInteractionResponseUpdateMessage {
+	type: InteractionResponseType.UpdateMessage;
+	data?: APIInteractionApplicationCommandCallbackData;
+}
+
+/**
+ * https://discord.com/developers/docs/interactions/slash-commands#interaction-response-interactionresponsetype
+ */
+export const enum InteractionResponseType {
+	/**
+	 * ACK a `Ping`
+	 */
+	Pong = 1,
+	/**
+	 * Respond to an interaction with a message
+	 */
+	ChannelMessageWithSource = 4,
+	/**
+	 * ACK an interaction and edit to a response later, the user sees a loading state
+	 */
+	DeferredChannelMessageWithSource,
+	/**
+	 * ACK a button interaction and update it to a loading state
+	 */
+	DeferredMessageUpdate,
+	/**
+	 * ACK a button interaction and edit the message to which the button was attached
+	 */
+	UpdateMessage,
+}
+
+/**
+ * https://discord.com/developers/docs/interactions/slash-commands#interaction-response-interactionapplicationcommandcallbackdata
+ */
+export type APIInteractionApplicationCommandCallbackData = Omit<
+	RESTPostAPIWebhookWithTokenJSONBody,
+	'username' | 'avatar_url'
+> & { flags?: MessageFlags };
